backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
redis[hiredis]==5.1.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.4
scipy==1.14.0
scikit-learn==1.5.0
torch==2.4.0
joblib==1.4.0
pydantic==2.9.0
